[tutorial pt 2] create models: Question and Choice
This commit defines database layout and additional
metadata for the Question and Choice classes.

A Question has a question_text field and a pub_date
field.

A Choice has choice_text field and a votes field
(in the form of an integer tally). Each Choice has
a Question foreign key.