[tutorial pt 3] update Polls index view to use template
The Polls index view function now has actual index
functionality: it displays an index of the 5 latest
poll questions in the system (making use of
Question.objects.order_by()).

Instead of hard-coding the HTML output in the view
itself, we've instructed the view function to load
a polls/index.html template. It also passes a 'context'
to the template in the form of a dictionary that maps
the 'latest_question_list' variable to the latest_
question_list collection returned by
Question.objects.order_by().

Django documentation's explanation of template
namespacing makes sense to me, but a filepath such
as 'polls/templates/polls/index.html' (or:
<app>/templates/<app>/template_file) still seems
convoluted and confusing.