[tutorial pt 3] update Polls index view to use render()
The index view now makes use of Django's render()
shortcut to:

- load the polls/index.html template
- pass a context to the template
- return an HttpResponse object

render() expects the following args: an HttpRequest
object, a template name, and a dictionary. It
returns an HttpResponse object of the given template
rendered with the given context.

Also - the README now includes a more detailed
explanation for the existence of this repo.