[tutorial pt 3] implement Polls namespace in URLconf
Django docs mention the problem with including
hardcoded links in templates, which makes it
challenging to change URLs on projects with lots
of templates. The docs recommend making use of
the fact that our polls.urls `urlpatterns` object
includes a `name` arg in the path() call for each
view, e.g.,:

path('<int:question_id>/', views.detail, name='detail')

Since we defined that name arg for each view, we
can take advantage of Django's {% url %} template
tag by replacing:
"/polls/{{ question.id }}/"

with:
"{% url 'detail' question.id %}"

I hesitated about embracing this piece of
automagic, but - why not.

And now that we're making use of the {% url %}
template tag, we should add a namespace to the
Polls URLconf. In the event that our project
includes another app that has a `detail` view, the
template tag is now able to locate the Polls detail
view thusly:

{% url 'polls:detail' question.id %}