[tutorial pt 4] update Polls vote view function
The "Vote" submit button on the Polls detail form
now has actual POST functionality: when a user
selects a radio button and submits the form, they
effectually submit a request including `choice=<id>`
data.

The form checks for a selected button and returns
an HttpResponseRedirect object if a selection was
made (i.e., the user is redirected to the results
view for the given question). We use
`request.POST['choice']` to check for the existence
of a POST request submission with 'choice' as its
key and a valid choice ID (PK) as its value.

Since the Polls vote view now serves as an example
of how to use Django's `get_object_or_404()`
shortcut, I removed the related comment from the
index view (which remains an example of how to
raise an Http404 exception without using the
shortcut).

I still don't completely understand the purpose of
using `reverse()` within the HttpResponseRedirect
constructor. Django docs explain that this has to
do with the avoidance of hardcoding a URL. I just
don't understand the exact role that `reverse()`
fulfills in supplying the URL string.

This commit also added a clarification to the README
regarding the difference between calling `get()` and
calling `filter()` on a collection of objects.