[tutorial pt 4] update Polls results view, add template
We now have a functional results view and template.
When a user submits a vote on a question choice,
they are redirected to a results page that shows
an actual vote tally for each choice (rather than
the generic stub view that we had before). They're
also invited to vote again via a link from the
results page.

NB: our vote() view is susceptible to a race condition.
It first retrieves the `selected_choice` object from
our database, then computes the new vote tally, and
then saves the new tally to the database.

The race condition lies in the possibility that two
users try to vote for the same choice at the exact same
time. Hypothetically, `selected_choice` would have an
existing vote tally of 42, and when both users attempt
to vote, a new tally of 43 would be computed and saved
(even though the truly accurate count would be 44).