[tutorial pt 5] add unit test: Question.was_published_recently()
We've identified a bug and added a unit test for the Question
model's `was_published_recently()` method, which returns True
even for questions that have a `pub_date` in the future.

In this commit, we've created a django.test.TestCase subclass
with a method that creates a Question instance with a future
`pub_date`. We then assert that `was_published_recently()`
returns False. (It currently returns True, so the test fails!)

This commit also corrects a typo in the Polls detail template
and adds a note/comment to the Polls index view function.